+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project changes will be shown here.
 - Issue 664 - Colour OS Support Dates
 - Issue 669 - Asset list extract remove duplicated logged in users names
 - Issue 661 - Review Meetings show primary active site
+- Issue 668 - Phone number cleansing
 ### Fixed
 - Issue 663 - Password Service Sorting not working